requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-admin-interface==0.12.2
django-admin-rangefilter==0.6.0
django-colorfield==0.3.0
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.1.0
django-rest-auth==0.9.5
django-summernote==0.8.11.6
djangorestframework==3.11.0
et-xmlfile==1.0.1
Faker==4.1.0
importlib-metadata==1.6.0
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.67
six==1.14.0
sqlparse==0.3.1
sweetify==2.0.0
tablib==2.0.0
text-unidecode==1.3
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.1.0
